se han corregido errores en la clase principal parcial2 y se ha hecho la clase magico.jva
